+++ src/jsMain/kotlin/app/seeds/Plan.kt
@@ -1,10 +1,14 @@
-import Plan.ComponentStyles.listDiv
+package app.seeds
+
+import app.TreeView
+import app.seeds.Plan.ComponentStyles.listDiv
 import com.ccfraser.muirwik.components.*
 import com.ccfraser.muirwik.components.button.mIconButton
 import com.ccfraser.muirwik.components.list.*
 import generated.model.SeedsDto
+import inputComponent
 import kotlinx.css.*
-//import models.Chore
+//import models.app.seeds.Chore
 import react.*
 import styled.StyleSheet
 import styled.css
